fixed bids-asks test and completed reducer test
